use sudo for create-dmg, rename "dmg" step to "createdmg", add a --debug-verbose option to enable extra-detailed output
The sudo for create-dmg seems to be needed to avoid a "permission denied" error in hdiutil on some hosts and configurations. I don't know exactly what triggers it. See issue 252 "hdiutil create fails with permission error..."

I like "createdmg" as a step name better, because it's a verb phrase like most of the other step names.