wip scripts: bump 9.1.0 to beta2, drop old lets-go-8-4-oh tap branch
The lets-go-8-4-oh branch is over now; that has been merged to main with the 8.4.0 GM release. I'm currently testing revised 9.x formulae on the WIP/polly-olly branch, which is for the C++ polymorphic allocator testing, but also has several unrelated fixes for the 9.x builds which will be used for regular releases, and just haven't been merged to main yet.

Bump 9.1.0 to beta2, since beta1 has been published