don't brew update/upgrade/cleanup, for stability and retaining cache
We no longer do the thing where you can upgrade some of the formulae in an in-progress octapp build in place. It's a one-shot thing, where it will (or should) use the versions that were in effect at the start of the build. That'll give us some more stability and reproducibility in the build, and avoid picking up detritus from upgraded formulae in the app. Plus, brew cleanup clears the downloads cache in the user directory, an we don't want that: we want to retain the cached downloads for use in subsequent build attempts, especially in cases with unreliable internet or flaky source hosting servers.