have custom pkg installation location use release suffixes like _beta1
This prevents users from accidentally re-using packages with native extensions built for a previous release and its baked-in path.

Only applies to prereleases and alternate builds, not to "u1" update suffixes. This is intentional; I think that updates should be back-compatible. Not sure though; may need to change that eventually.  Leaning towards not doing it yet, because the updates are suppsoed to be pretty drop-in replacements, and ones more likely to be used by end users (as opposed by devs or power users testing betas), so it's a bigger burden to ask them to install all their packages again.