make GUI and desktop non-optional
Octave.app is fundamentally a GUI app, so doing a CLI-based build of it doesn't make much sense, and I haven't done so in years. If you really want CLI stuff that makes use of the Octave.app customizations, you can use the formulae in the octave-app/octave-app custom tap to install those into your Homebrew installation and use them there. Octave.app is never going to be distributing an installer of a non-GUI build, so I don't think we need to support it in the app bundler any more.