remove --no-java option, controlling this with formulae instead
The use or non-use of Java is now part of the formula definitions. We'll likely never want to publish a non-Java build of Octave.app, and if we do, that should be built in to formula and release definition metadata, instead of being controlled by a manual `bundle_octapp` option. So just remove this option.