remove logger use by app launcher; syslog is wrong and it prob doesn't work anyway
I don't think user apps like Octave should generally be sending messages to syslog; that's spammy, and Octave in particular doesn't really define what it does with its stdout/stderr stuff. If we're logging stuff, I think it should go to per-app-execution log files under /var, /tmp, or maybe some user log, and probably one that's ephemeral or rotated, because it could be spammy; that's a lot of executions.

Plus, I don't think this was working anyway. Because of permissions and security, I don't think regular users can write to syslog using `logger`. And I don't know if logger even accepts a stdin stream; I think it takes its message as a command argument. So this maybe never worked in the first place.